move-PlatformParker-out-of-os-xxx-hpp
