+++ src/hotspot/os/posix/os_posix.hpp
@@ -97,65 +97,6 @@ public:
 };
 
 
-/*
- * This is the platform-specific implementation underpinning
- * the ParkEvent class, which itself underpins Java-level monitor
- * operations. See park.hpp for details.
- * These event objects are type-stable and immortal - we never delete them.
- * Events are associated with a thread for the lifetime of the thread.
- */
-class PlatformEvent : public CHeapObj<mtSynchronizer> {
- private:
-  double cachePad[4];        // Increase odds that _mutex is sole occupant of cache line
-  volatile int _event;       // Event count/permit: -1, 0 or 1
-  volatile int _nParked;     // Indicates if associated thread is blocked: 0 or 1
-  pthread_mutex_t _mutex[1]; // Native mutex for locking
-  pthread_cond_t  _cond[1];  // Native condition variable for blocking
-  double postPad[2];
-
- protected:       // TODO-FIXME: make dtor private
-  ~PlatformEvent() { guarantee(false, "invariant"); } // immortal so can't delete
-
- public:
-  PlatformEvent();
-  void park();
-  int  park(jlong millis);
-  void unpark();
-
-  // Use caution with reset() and fired() -- they may require MEMBARs
-  void reset() { _event = 0; }
-  int  fired() { return _event; }
-};
-
-// JSR166 support
-// PlatformParker provides the platform dependent base class for the
-// Parker class. It basically provides the internal data structures:
-// - mutex and convars
-// which are then used directly by the Parker methods defined in the OS
-// specific implementation files.
-// There is significant overlap between the funcionality supported in the
-// combination of Parker+PlatformParker and PlatformEvent (above). If Parker
-// were more like ObjectMonitor we could use PlatformEvent in both (with some
-// API updates of course). But Parker methods use fastpaths that break that
-// level of encapsulation - so combining the two remains a future project.
-
-class PlatformParker {
-  NONCOPYABLE(PlatformParker);
- protected:
-  enum {
-    REL_INDEX = 0,
-    ABS_INDEX = 1
-  };
-  volatile int _counter;
-  int _cur_index;  // which cond is in use: -1, 0, 1
-  pthread_mutex_t _mutex[1];
-  pthread_cond_t  _cond[2]; // one for relative times and one for absolute
-
- public:
-  PlatformParker();
-  ~PlatformParker();
-};
-
 // Workaround for a bug in macOSX kernel's pthread support (fixed in Mojave?).
 // Avoid ever allocating a pthread_mutex_t at the same address as one of our
 // former pthread_cond_t, by using freelists of mutexes and condvars.
