+++ src/hotspot/share/runtime/park.hpp
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 1997, 2021, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 1997, 2022, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -28,6 +28,16 @@
 #include "utilities/debug.hpp"
 #include "utilities/globalDefinitions.hpp"
 
+#if defined(LINUX) || defined(AIX) || defined(BSD)
+# include "park_posix.hpp"
+#elif defined(_WINDOWS)
+# include "park_windows.hpp"
+#else
+# error "No PlatformParker implementation provided for this OS"
+#endif
+
+class Thread;
+
 /*
  * Per-thread blocking support for JSR166. See the Java-level
  * documentation for rationale. Basically, park acts like wait, unpark
@@ -39,7 +49,7 @@
  * ThreadsListHandle.
  *
  * Class Parker is declared in shared code and extends the platform-specific
- * os::PlatformParker class, which contains the actual implementation
+ * PlatformParker class, which contains the actual implementation
  * mechanics (condvars/events etc). The implementation for park() and unpark()
  * are also in the platform-specific os_<os>.cpp files.
  *
@@ -49,7 +59,7 @@
  *
  */
 
-class Parker : public os::PlatformParker {
+class Parker : public PlatformParker {
  private:
   NONCOPYABLE(Parker);
  public:
@@ -102,7 +112,7 @@ class Parker : public os::PlatformParker {
 // We'll want to eventually merge these redundant facilities and use ParkEvent.
 
 
-class ParkEvent : public os::PlatformEvent {
+class ParkEvent : public PlatformEvent {
   private:
     ParkEvent * FreeNext ;
